ch04_similarity_search/requirements.txt
numpy==2.2.6
faiss-cpu==1.12.0
rich==14.2.0
pytest==8.3.5
